pyproject.toml
[tool.poetry]
name = "learn_ai"
version = "0.0.1"
description = "FuturePath AI's training course on Generative AI"
authors = ["Maruti Agarwal <maruti@futurepath.ai>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
scikit-learn = "^1.2.2, <=1.3.2"
scipy = "^1.10.1"
easydict = "^1.10"
xgboost = "^1.7.5"
ipywidgets = "^8.0.6"
tabulate = "^0.9.0"
requests = "^2.28.2"
numpy = "^1.23.5"
jupyterlab = "^4.0.5"
openai = "*"
tqdm = "^4.66.1"
pandas = "^2.1.2"
langchain = "*"
unstructured = {extras = ["pdf"], version = "^0.12.2"}
tiktoken = "^0.5.1"
sentence-transformers = "^2.2.2"
matplotlib = "^3.8.1"
llama-index = "^0.9.28.post1"
ragas = "^0.0.22"
chromadb = "^0.4.22"
langchain-openai = "^0.0.3"
outlines = "^0.0.24"
xmltodict = "^0.13.0"
termcolor = "^2.4.0"
tenacity = "^8.2.3"
python-dotenv = "^1.0.1"
voyager = "^2.0.2"
html5lib = "^1.1"
instructor = {version = "^0.6.1", python = ">=3.10,<3.11"}

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"
graphviz = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
